// pro/src/apiClient/hey-api/types.gen.ts
export type PokeTodoPriority = 'low' | 'medium' | 'high'

/**
 * PokeTodoResponseModel
 */
export type PokeTodoResponseModel = {
  /**
   * Datecreated
   */
  dateCreated: string
  /**
   * Dateupdated
   */
  dateUpdated: string | null
  /**
   * Description
   */
  description: string | null
  /**
   * Duedate
   */
  dueDate: string | null
  /**
   * Id
   */
  id: number
  /**
   * Iscompleted
   */
  isCompleted: boolean
  priority: PokeTodoPriority
  /**
   * Title
   */
  title: string
}